refactor: remove LessonBlock component and use lesson blocks directly
