--- front/src/components/LessonBlock/index.js
@@ -1,8 +1,0 @@
-import LessonBlock, {
-  ONGOING_SHORT_LESSON,
-  ONGOING_FULL_LESSON,
-  PUBLIC_LESSON,
-} from './LessonBlock';
-
-export { ONGOING_SHORT_LESSON, ONGOING_FULL_LESSON, PUBLIC_LESSON };
-export default LessonBlock;
