+++ front/src/components/lessonBlocks/OngoingFull/OngoingFull.desktop.jsx
