+++ front/src/components/lessonBlocks/OngoingFull/OngoingFull.desktop.styled.js
