+++ front/src/components/lessonBlocks/OngoingFull/OngoingFull.mobile.jsx
