+++ front/src/components/lessonBlocks/OngoingFull/OngoingFull.mobile.styled.js
