+++ front/src/components/lessonBlocks/OngoingShort/OngoingShort.mobile.jsx
