+++ front/src/components/lessonBlocks/OngoingShort/OngoingShort.styled.js
