+++ front/src/components/lessonBlocks/Public/Public.desktop.jsx
