+++ front/src/components/lessonBlocks/Public/Public.mobile.jsx
