+++ front/src/components/lessonBlocks/Public/Public.mobile.styled.js
