+++ front/src/pages/Teacher/Home/LessonsDashboard/LessonsList/LessonsList.jsx
@@ -1,6 +1,5 @@
 import PropTypes from 'prop-types';
-import LessonBlock from '@sb-ui/components/LessonBlock';
-import { TEACHER_LESSON } from '@sb-ui/components/LessonBlock/LessonBlock';
+import TeacherLesson from '@sb-ui/components/lessonBlocks/Teacher';
 import AddCard from '../AddCard';
 
 import { CardCol } from '../LessonsDashboard.styled';
@@ -14,8 +13,7 @@ const LessonsList = ({ lessons, onCreateLesson, isAddNewShown }) => (
     ) : (
       lessons.map((lesson) => (
         <CardCol key={lesson.id} span={12}>
-          <LessonBlock
-            type={TEACHER_LESSON}
+          <TeacherLesson
             id={lesson.id}
             title={lesson.name}
             students={lesson.students}
