+++ front/src/pages/User/Home/OngoingLessons/LessonsList/LessonsList.jsx
@@ -3,9 +3,7 @@ import PropTypes from 'prop-types';
 import { useTranslation } from 'react-i18next';
 import emptyImg from '@sb-ui/resources/img/empty.svg';
 import useMobile from '@sb-ui/hooks/useMobile';
-import LessonBlock, {
-  ONGOING_SHORT_LESSON,
-} from '@sb-ui/components/LessonBlock';
+import OngoingShortLesson from '@sb-ui/components/lessonBlocks/OngoingShort';
 import * as S from './LessonsList.styled';
 
 const LessonsList = ({ lessons, isLoading }) => {
@@ -30,7 +28,7 @@ const LessonsList = ({ lessons, isLoading }) => {
           <>
             {lessons?.map((lesson) => (
               <S.LessonsColumn xl={{ span: 8 }} lg={{ span: 24 }}>
-                <LessonBlock type={ONGOING_SHORT_LESSON} lesson={lesson} />
+                <OngoingShortLesson lesson={lesson} />
               </S.LessonsColumn>
             ))}
           </>
